code separation into controllers and routes
